Change decoder output to float32 for mixed precision training

+++ dall_e_tf/decoder.py
@@ -29,5 +29,5 @@ def dvae_decoder(img_size=None, group_count=4, n_init=128, n_hid=256, n_blk_per_
         if i > 0:
             x = UpSampling2D(2)(x)
     x = act_fn(x)
-    x = Conv2D(output_channels, 1)(x) # Todo: figure out why this is 2x output channels
+    x = Conv2D(output_channels, 1, dtype=tf.float32)(x) # Todo: figure out why this is 2x output channels
     return Model(x_in, x)
